Fix handling of common indent with non-matching lines
If the range to be aligned has non-matching lines, they still affect how
the common indent is determined.  This is most noticeable when there are
blank lines in the range as all the lines which get aligned are thrown
to column 0.

LongestCommonIndent now ignores any whitespace only lines and, when
:GTabularize is called, any lines not matching the delimiter.

The stripping and restoring of the common indent is also more selective
to avoid restoring the indent to lines which weren't aligned.

Signed-off-by: James McCoy <vega.james@gmail.com>